Fjerner gammel breadcrumbs-kode
