--- src/components/breadcrumbs/Breadcrumbs.less
@@ -1,32 +1,0 @@
-
-.breadcrumbs {
-    display: flex;
-    flex-direction: row;
-    flex-wrap: wrap;
-    align-items: center;
-
-    &__ikon {
-      width: 24px;
-      height: 24px;
-      margin-right: 0.5em;
-    }
-
-    &__chevron {
-        color: @navGra60;
-        display: inline-block;
-        margin: 0 0.5em;
-    }
-
-    &__segment {
-        white-space: nowrap;
-        display: inline-block;
-
-        &:last-child {
-            font-weight: bold;
-        }
-    }
-
-    @media (max-width: @screen-tablet) {
-        align-self: flex-start;
-    }
-}
