--- src/components/topp-linje/ToppLinje.less
@@ -1,14 +1,0 @@
-
-.kontakt-oss-topplinje {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  align-items: center;
-  width: 100%;
-  padding: 0 1rem;
-  box-sizing: border-box;
-
-  @media (max-width: @screen-tablet) {
-    flex-direction: column;
-  }
-}
