+++ src/index.less
@@ -36,8 +36,6 @@
 @import "components/box/Box.less";
 @import "components/icon/Icon.less";
 
-@import "components/breadcrumbs/Breadcrumbs.less";
-
 @import "components/ikonpanel/IkonPanel.less";
 @import "components/routerlenke/RouterLenkeMedChevron.less";
 @import "components/routerlenke/RouterLenke.less";
@@ -47,5 +45,3 @@
 @import "components/varsler/korona-virus-varsel/KoronaVirusVarsel.less";
 @import "components/varsler/Varsler.less";
 @import "components/content-loader/NavContentLoader.less";
-
-@import "components/topp-linje/ToppLinje.less";
